Calculate Returns From Stock Prices

Stock prices are known for been non-stationary. Use get_returns() as a pre-processing step for econometric modeling.

get_returns(.tbl, .group, .type, .omit_na = FALSE, ...)

# S3 method for default
get_returns(.tbl, .group, .type, .omit_na = FALSE, ...)

# S3 method for tbl_df
get_returns(.tbl, .group, .type, .omit_na = FALSE, ...)

Arguments

.tbl
A tidy tibble.

.group
The column in which the data should be grouped. This will often be a column with stock tickers or stocks names.

.type
The method used to calculate returns. One of:

log

arithmetic

.omit_na
Should NA values be omitted? Default is FALSE (NA's are replaced by 0).

...
The column (or columns) in which the calculation should be conducted.

Value

A tidy tibble.

Examples

library(YahooTickers)
library(dplyr)
get_tickers(dow) %>%
  slice(1:3) %>%
  get_stocks(.) %>%
  get_returns(., .group = tickers, .type = arithmetic,
              .omit_na = TRUE, adjusted)

#> # A tibble: 8,286 x 3
#>    date       tickers adjusted
#>    <date>     <fct>      <dbl>
#>  1 2010-01-05 HD       0.00732
#>  2 2010-01-06 HD      -0.00346
#>  3 2010-01-07 HD       0.0118 
#>  4 2010-01-08 HD      -0.00481
#>  5 2010-01-11 HD      -0.0283 
#>  6 2010-01-12 HD      -0.00639
#>  7 2010-01-13 HD       0.00536
#>  8 2010-01-14 HD       0.00604
#>  9 2010-01-15 HD       0.00954
#> 10 2010-01-19 HD       0.0109 
#> # … with 8,276 more rows