A wrapper over getSymbols that maps over index components.

get_stocks() allows you to quickly download stock indexes from the the Yahoo Finance without having to run for loops.

get_stocks(
  tickers,
  from = "2010-01-01",
  to = Sys.Date(),
  periodicity = "daily",
  simplify = TRUE,
  otherwise = NULL,
  quiet = TRUE
)

Arguments

tickers
A character vector with the stock tickers.

from
A date in the format YYYY-MM-DD.

to
A date in the format YYYY-MM-DD.

periodicity
One of: daily, weely or monthly.

simplify
If TRUE, a tibble is returned. If FALSE a list with two elements is returned:

result: with the downloads that succeeded;

error: with the stocks not covered by the Yahoo Finance.

The defaul is TRUE.

otherwise
Argument passed to safely.

quiet
Argument passed to safely. A notification is showed whenever an error occurs. Default is TRUE.

Value

A tidy tibble if simplify = TRUE and a list of two components if simplify = FALSE.

See also

Examples

# Get stocks from the Merval Stock Index (Argentina)
library(YahooTickers)
ticks <- get_tickers(merval)
get_stocks(ticks, periodicity = "monthly")

#> Warning: PESA.BA download failed; trying again.

#> Warning: TS.BA download failed; trying again.

#> Warning: APBR.BA download failed; trying again.

#> Warning: ERAR.BA download failed; trying again.

#> Warning: FRAN.BA download failed; trying again.

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> Warning: `type_convert()` only converts columns of type 'character'.
#> - `df` has no columns of type 'character'

#> # A tibble: 1,136 x 8
#>    date       tickers  open  high   low close   volume adjusted
#>    <date>     <fct>   <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 2010-01-01 COME.BA 0.131 0.155 0.131 0.143 23477234    0.121
#>  2 2010-02-01 COME.BA 0.144 0.144 0.126 0.128 13070731    0.108
#>  3 2010-03-01 COME.BA 0.131 0.220 0.126 0.184 59390001    0.156
#>  4 2010-04-01 COME.BA 0.184 0.220 0.170 0.202 45498717    0.171
#>  5 2010-05-01 COME.BA 0.202 0.211 0.163 0.190 33671218    0.161
#>  6 2010-06-01 COME.BA 0.187 0.192 0.166 0.166 14034908    0.141
#>  7 2010-07-01 COME.BA 0.167 0.183 0.159 0.175 15794723    0.148
#>  8 2010-08-01 COME.BA 0.176 0.177 0.157 0.159 12933754    0.134
#>  9 2010-09-01 COME.BA 0.160 0.207 0.158 0.182 29222468    0.154
#> 10 2010-10-01 COME.BA 0.182 0.232 0.179 0.224 49571056    0.189
#> # ... with 1,126 more rows