A wrapper over getSymbols that maps over index components.

get_stocks() allows you to quickly download stock indexes from the the Yahoo Finance without having to run for loops.

get_stocks(
  tickers,
  from = "2010-01-01",
  to = Sys.Date(),
  periodicity = "daily",
  simplify = TRUE,
  otherwise = NULL,
  quiet = TRUE
)

Arguments

tickers
A character vector with the stock tickers.

from
A date in the format YYYY-MM-DD.

to
A date in the format YYYY-MM-DD.

periodicity
One of: daily, weely or monthly.

simplify
If TRUE, a tibble is returned. If FALSE a list with two elements is returned:

result: with the downloads that succeeded;

error: with the stocks not covered by the Yahoo Finance.

The defaul is TRUE.

otherwise
Argument passed to safely.

quiet
Argument passed to safely. A notification is showed whenever an error occurs. Default is TRUE.

Value

A tidy tibble if simplify = TRUE and a list of two components if simplify = FALSE.

See also

Examples

# Get stocks from the Merval Stock Index (Argentina)
library(YahooTickers)
ticks <- get_tickers(merval)
get_stocks(ticks, periodicity = "monthly")

#> Warning: Problem with `mutate()` input `download`.
#> i PESA.BA download failed; trying again.
#> i Input `download` is `purrr::map(...)`.

#> Warning: PESA.BA download failed; trying again.

#> Warning: Problem with `mutate()` input `download`.
#> i TS.BA download failed; trying again.
#> i Input `download` is `purrr::map(...)`.

#> Warning: TS.BA download failed; trying again.

#> Warning: Problem with `mutate()` input `download`.
#> i APBR.BA download failed; trying again.
#> i Input `download` is `purrr::map(...)`.

#> Warning: APBR.BA download failed; trying again.

#> Warning: Problem with `mutate()` input `download`.
#> i ERAR.BA download failed; trying again.
#> i Input `download` is `purrr::map(...)`.

#> Warning: ERAR.BA download failed; trying again.

#> Warning: Problem with `mutate()` input `download`.
#> i FRAN.BA download failed; trying again.
#> i Input `download` is `purrr::map(...)`.

#> Warning: FRAN.BA download failed; trying again.

#> # A tibble: 1,056 x 8
#>    date       tickers  open  high   low close   volume adjusted
#>    <date>     <fct>   <dbl> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 2010-01-01 EDN.BA   1.54  1.61  1.38  1.44 12068374     1.44
#>  2 2010-02-01 EDN.BA   1.44  1.46  1.31  1.34  8019848     1.34
#>  3 2010-03-01 EDN.BA   1.34  1.48  1.31  1.42  9077880     1.42
#>  4 2010-04-01 EDN.BA   1.42  1.54  1.41  1.45  7609308     1.45
#>  5 2010-05-01 EDN.BA   1.45  1.54  1.38  1.42  7911524     1.42
#>  6 2010-06-01 EDN.BA   1.42  1.45  1.27  1.29  7943458     1.29
#>  7 2010-07-01 EDN.BA   1.27  1.43  1.26  1.41 11323493     1.41
#>  8 2010-08-01 EDN.BA   1.42  1.42  1.25  1.27  8881487     1.27
#>  9 2010-09-01 EDN.BA   1.29  1.62  1.28  1.56 13344430     1.56
#> 10 2010-10-01 EDN.BA   1.56  2.06  1.53  2    19326303     2   
#> # ... with 1,046 more rows